김기현 - groupchat.js 기능별로 파일 분리 -> /assets/js/groupchat/init.js
