+++ src/main/resources/application.properties
@@ -1,0 +1,1 @@
+file.upload-dir=C:/upload/images/group
