+++ src/main/webapp/WEB-INF/config/application-context.xml
@@ -19,4 +19,5 @@
 		p:suffix=".jsp"
 	/>
 	<bean id="multipartResolver" class="org.springframework.web.multipart.commons.CommonsMultipartResolver"/>
+	<context:property-placeholder location="classpath:application.properties" />
 </beans>
